yearly view works - wip
