+++ src/app/domains/statistics/statistics-filter.component.ts
@@ -62,7 +62,17 @@ export class StatisticFilterComponent implements OnInit {
     private translateService: TranslateService,
     private statisticsFilterService: StatisticsFilterService,
     private fb: FormBuilder
-  ) {}
+  ) {
+    this.filter = this.fb.group({
+      year: [],
+      datasource: [],
+      analyticsType: [],
+      species: [],
+      phenophase: [],
+      altitude: [],
+      graph: []
+    });
+  }
 
   ngOnInit() {
     // fixme todo check if needed
@@ -77,22 +87,30 @@ export class StatisticFilterComponent implements OnInit {
     // bind filter values to form
     this.subscriptions.add(
       this.statisticsFilterService.currentFilters$.subscribe(cf => {
-        this.filter = this.fb.group({
-          year: [cf.year],
-          datasource: [cf.datasource],
-          analyticsType: [cf.analyticsType],
-          species: [cf.species],
-          phenophase: [cf.phenophase],
-          altitude: [cf.altitude],
-          graph: [cf.graph]
+        console.log('update filter values', cf);
+
+        this.filter.setValue({
+          year: cf.year,
+          datasource: cf.datasource,
+          analyticsType: cf.analyticsType,
+          species: cf.species,
+          phenophase: cf.phenophase,
+          altitude: cf.altitude,
+          graph: cf.graph
         });
+      })
+    );
+
+    this.subscriptions.add(
+      this.filter.valueChanges.subscribe(value => {
+        this.statisticsFilterService.updateFilters(value);
       })
     );
 
     // dynamic filters
     this.selectableYears$ = this.statisticsFilterService.getSelectableYears();
     this.selectableSources$ = of(sourceFilterValues);
-    this.selectableAnalyticsTypes$ = of(selectableAnalyticsTypes);
+    this.selectableAnalyticsTypes$ = this.statisticsFilterService.getSelectableAnalyticsTypes();
     this.selectableSpecies$ = this.statisticsFilterService.getSelectableSpecies();
     this.selectablePhenophases$ = this.statisticsFilterService.getSelectablePhenophases();
     this.selectableAltitudeGroups$ = of(selectableAltitudeGroup);
