+++ src/app/domains/statistics/yearly-statistics.component.html
@@ -1,1 +1,3 @@
-<p>yearly-statistics works!</p>
+<div #statisticsContainer>
+  <svg id="statistics-graph" width="100%" attr.height="{{ svgComponentHeight }}px"></svg>
+</div>
